Refactor and modernize datatools/configuration/io classes
Split out ascii_io, variant_preprocessor, and variant_reporting classes
into separate headers/srcs, one per class.

Retain datatools/configuration/io.h header, simply including the
three new headers for interface compatibility.

Modernize variant io classes following ISO C++ Guidelines

- Use range based for where possible
- Use default/in class construction
- Normal order member functions

Add a constructor to variant_preprocessor taking the repository
to be used along with the flags.

This allows clients to avoid using the datatools kernel, and follow
"construction is initialization".
